+++ test/core/eventify/test.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="utf-8">
         <title>Future.js Test - Eventify</title>
-        <link rel="stylesheet" type="text/css" href="../../node_modules/qunitjs/qunit/qunit.css">
+        <link rel="stylesheet" type="text/css" href="../../../node_modules/qunitjs/qunit/qunit.css">
     </head>
     <body>
         <div id="qunit"></div>
