+++ test/core/eventify/test.js
@@ -1,6 +1,6 @@
-import '../sauce.js';
+import '../../sauce.js';
 import QUnit from 'qunitjs';
-import eventify from '../../src/core/eventify.js';
+import eventify from '../../../src/core/eventify.js';
 
 
 QUnit.module('eventify');
